Chore: Remove pluginId references
