+++ packages/plugins/users-permissions/admin/src/pages/Roles/components/ProtectedCreatePage.js
@@ -2,14 +2,11 @@ import React from 'react';
 
 import { CheckPagePermissions } from '@strapi/helper-plugin';
 
-import { PERMISSIONS } from '../../constants';
+import { PERMISSIONS } from '../../../constants';
+import { CreatePage } from '../pages/CreatePage';
 
-import { CreatePage } from './CreatePage';
-
-const ProtectedRolesCreatePage = () => (
+export const ProtectedRolesCreatePage = () => (
   <CheckPagePermissions permissions={PERMISSIONS.createRole}>
     <CreatePage />
   </CheckPagePermissions>
 );
-
-export default ProtectedRolesCreatePage;
