+++ packages/plugins/users-permissions/admin/src/pages/Roles/index.js
@@ -4,23 +4,26 @@ import { AnErrorOccurred, CheckPagePermissions } from '@strapi/helper-plugin';
 import { Route, Switch } from 'react-router-dom';
 
 import { PERMISSIONS } from '../../constants';
-import pluginId from '../../pluginId';
 
-import ProtectedRolesCreatePage from './ProtectedCreatePage';
-import ProtectedRolesEditPage from './ProtectedEditPage';
-import ProtectedRolesListPage from './ProtectedListPage';
+import { ProtectedRolesCreatePage } from './components/ProtectedCreatePage';
+import { ProtectedRolesEditPage } from './components/ProtectedEditPage';
+import { ProtectedRolesListPage } from './components/ProtectedListPage';
 
 const Roles = () => {
   return (
     <CheckPagePermissions permissions={PERMISSIONS.accessRoles}>
       <Switch>
         <Route
-          path={`/settings/${pluginId}/roles/new`}
+          path="/settings/users-permissions/roles/new"
           component={ProtectedRolesCreatePage}
           exact
         />
-        <Route path={`/settings/${pluginId}/roles/:id`} component={ProtectedRolesEditPage} exact />
-        <Route path={`/settings/${pluginId}/roles`} component={ProtectedRolesListPage} exact />
+        <Route
+          path="/settings/users-permissions/roles/:id"
+          component={ProtectedRolesEditPage}
+          exact
+        />
+        <Route path="/settings/users-permissions/roles" component={ProtectedRolesListPage} exact />
         <Route path="" component={AnErrorOccurred} />
       </Switch>
     </CheckPagePermissions>
