+++ packages/plugins/users-permissions/admin/src/pages/Roles/pages/CreatePage.js
@@ -1,76 +1,78 @@
 import * as React from 'react';
 
 import {
+  Button,
   ContentLayout,
+  Flex,
+  Grid,
+  GridItem,
   HeaderLayout,
   Main,
-  Button,
-  Flex,
-  TextInput,
   Textarea,
+  TextInput,
   Typography,
-  GridItem,
-  Grid,
 } from '@strapi/design-system';
 import {
-  useFetchClient,
-  useOverlayBlocker,
-  SettingsPageTitle,
-  LoadingIndicatorPage,
   Form,
-  useFormatAPIError,
+  SettingsPageTitle,
+  useFetchClient,
   useNotification,
-  Link,
+  useOverlayBlocker,
+  useTracking,
 } from '@strapi/helper-plugin';
-import { ArrowLeft, Check } from '@strapi/icons';
+import { Check } from '@strapi/icons';
 import { Formik } from 'formik';
 import { useIntl } from 'react-intl';
 import { useMutation } from 'react-query';
-import { useRouteMatch } from 'react-router-dom';
+import { useHistory } from 'react-router-dom';
 
-import UsersPermissions from '../../components/UsersPermissions';
-import { usePlugins, useFetchRole } from '../../hooks';
-import getTrad from '../../utils/getTrad';
+import UsersPermissions from '../../../components/UsersPermissions';
+import { usePlugins } from '../../../hooks';
+import getTrad from '../../../utils/getTrad';
+import { createRoleSchema } from '../constants';
 
-import { createRoleSchema } from './constants';
-
-export const EditPage = () => {
+export const CreatePage = () => {
   const { formatMessage } = useIntl();
   const toggleNotification = useNotification();
+  const { goBack } = useHistory();
   const { lockApp, unlockApp } = useOverlayBlocker();
-  const {
-    params: { id },
-  } = useRouteMatch(`/settings/users-permissions/roles/:id`);
-  const { isLoading: isLoadingPlugins, routes } = usePlugins();
-  const { role, onSubmitSucceeded, isLoading: isLoadingRole } = useFetchRole(id);
+  const { isLoading: isLoadingPlugins, permissions, routes } = usePlugins();
+  const { trackUsage } = useTracking();
   const permissionsRef = React.useRef();
-  const { put } = useFetchClient();
-  const { formatAPIError } = useFormatAPIError();
-  const mutation = useMutation((body) => put(`/users-permissions/roles/${id}`, body), {
-    onError(error) {
+  const { post } = useFetchClient();
+  const mutation = useMutation((body) => post(`/users-permissions/roles`, body), {
+    onError() {
       toggleNotification({
         type: 'warning',
-        message: formatAPIError(error),
+        message: {
+          id: 'notification.error',
+          defaultMessage: 'An error occurred',
+        },
       });
     },
 
-    onSuccess(data) {
+    onSuccess() {
+      trackUsage('didCreateRole');
+
       toggleNotification({
         type: 'success',
         message: {
           id: getTrad('Settings.roles.created'),
-          defaultMessage: 'Role edited',
+          defaultMessage: 'Role created',
         },
       });
 
-      onSubmitSucceeded({ name: data.name, description: data.description });
+      // Forcing redirecting since we don't have the id in the response
+      goBack();
     },
   });
 
-  const handleEditRoleSubmit = async (data) => {
-    // Set loading state
+  const handleCreateRoleSubmit = async (data) => {
     lockApp();
 
+    // TODO: refactor. Child -> parent component communication is evil;
+    // We should either move the provider one level up or move the state
+    // straight into redux.
     const permissions = permissionsRef.current.getPermissions();
 
     await mutation.mutate({ ...data, ...permissions, users: [] });
@@ -78,18 +80,14 @@ export const EditPage = () => {
     unlockApp();
   };
 
-  if (isLoadingRole) {
-    return <LoadingIndicatorPage />;
-  }
-
   return (
     <Main>
-      {/* TODO: this needs to be translated */}
+      {/* TODO: This needs to be translated */}
       <SettingsPageTitle name="Roles" />
       <Formik
         enableReinitialize
-        initialValues={{ name: role.name, description: role.description }}
-        onSubmit={handleEditRoleSubmit}
+        initialValues={{ name: '', description: '' }}
+        onSubmit={handleCreateRoleSubmit}
         validationSchema={createRoleSchema}
       >
         {({ handleSubmit, values, handleChange, errors }) => (
@@ -97,12 +95,7 @@ export const EditPage = () => {
             <HeaderLayout
               primaryAction={
                 !isLoadingPlugins && (
-                  <Button
-                    disabled={role.code === 'strapi-super-admin'}
-                    type="submit"
-                    loading={mutation.isLoading}
-                    startIcon={<Check />}
-                  >
+                  <Button type="submit" loading={mutation.isLoading} startIcon={<Check />}>
                     {formatMessage({
                       id: 'global.save',
                       defaultMessage: 'Save',
@@ -110,16 +103,14 @@ export const EditPage = () => {
                   </Button>
                 )
               }
-              title={role.name}
-              subtitle={role.description}
-              navigationAction={
-                <Link startIcon={<ArrowLeft />} to="/settings/users-permissions/roles">
-                  {formatMessage({
-                    id: 'global.back',
-                    defaultMessage: 'Back',
-                  })}
-                </Link>
-              }
+              title={formatMessage({
+                id: 'Settings.roles.create.title',
+                defaultMessage: 'Create a role',
+              })}
+              subtitle={formatMessage({
+                id: 'Settings.roles.create.description',
+                defaultMessage: 'Define the rights given to the role',
+              })}
             />
             <ContentLayout>
               <Flex
@@ -134,7 +125,7 @@ export const EditPage = () => {
                 paddingRight={7}
                 shadow="filterShadow"
               >
-                <Flex direction="column" alignItems="stretch" gap={4}>
+                <Flex direction="column" alignItems="stretch">
                   <Typography variant="delta" as="h2">
                     {formatMessage({
                       id: getTrad('EditPage.form.roles'),
@@ -173,7 +164,7 @@ export const EditPage = () => {
                 {!isLoadingPlugins && (
                   <UsersPermissions
                     ref={permissionsRef}
-                    permissions={role.permissions}
+                    permissions={permissions}
                     routes={routes}
                   />
                 )}
